+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.38.2-0 / Unreleased
+  * [Fixed] #1097 Removed incorrect set-function for jQuery cookie
 
 # 2.38.1 / 2022-01-14
   * Replace deprecated jQuery functions: `$.isArray()`, `$.isFunction()`,