+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# 2.38.2-0 / Unreleased
+# 2.38.3-0 / Unreleased
+  * #1113 Update jQuery UI to 1.13.2
+
+# 2.38.2-0 / 2022-06-30
   * [Fixed] #1097 Removed incorrect set-function for jQuery cookie
   * [Fixed] #1108 minExpandLevel: 2 together with table extension does not work